home completed

+++ admin/src/api/customers.api.js
@@ -1,10 +1,10 @@
 import api from "./axios";
 
 export const getAllCustomersApi = (params) =>
-    api.get("/customers", { params });
+    api.get("/admin-customers", { params });
 
 export const updateCustomerStatusApi = (id, status) =>
-    api.patch(`/customers/${id}/status`, { status });
+    api.patch(`/admin-customers/${id}/status`, { status });
 
 export const deleteCustomerApi = (id) =>
-    api.delete(`/customers/${id}`);
+    api.delete(`/admin-customers/${id}`);
